cfgfile: Squelch a last compiler warning
We do this by casting the one location where we referred to an
__off64_t in printing to a (long long) unconditionally and then
using the %lld print format specifier in the format string.

Since it'll mean we're casting to a larger type on 64-bit archs
and the same width on 32-bit ones, we needn't worry about losses.

+++ cfgfile.c
@@ -43,8 +43,8 @@ static char *cfg_read_file(const char *path, unsigned *len)
 	buf = malloc(st.st_size + 3);
 	if (!buf) {
 		*len = -errno;
-		fprintf(stderr, "Failed to allocate %lu bytes of memory for '%s'\n",
-				st.st_size, path);
+		fprintf(stderr, "Failed to allocate %lld bytes of memory for '%s'\n",
+		        (long long)st.st_size, path);
 		close(fd);
 		return NULL;
 	}
